+++ web/ts/data-store/data-store.ts
@@ -1,5 +1,7 @@
 import {VideoSectionProvider} from "./video-sections";
+import {BookmarksProvider} from "./bookmarks";
 
 export abstract class DataStore {
+    static bookmarks : BookmarksProvider = new BookmarksProvider();
     static videoSections : VideoSectionProvider = new VideoSectionProvider();
 }
